New Correction

+++ 0x00-hello_world/6-size.c
@@ -15,7 +15,7 @@ long long alonglong;
 float afloat;
 
 printf("Size of a char: %zu byte(s)\n", (unsigned long) sizeof(acharacter));
-printf("size of an int: %zu byte(s)\n", (unsigned long) sizeof(ainteger));
+printf("size of an int: %zu byte(s)\n", sizeof(ainteger));
 printf("Size of a long int: %zu byte(s)\n", (unsigned long) sizeof(along));
 printf("Size of a long long int: %zu byte(s)\n", (unsigned long) sizeof(alonglong));
 printf("Size of a float: %zu byte(s)\n", (unsigned long) sizeof(afloat));
